Renew wasm interpreter

+++ apps/bball/src/main.rs
@@ -13,8 +13,8 @@ fn _start() {
         for (j, t2) in TABLE[i..].iter().enumerate() {
             let dis = if j < 8 { j } else { 15 - j };
             window.draw_line(
-                Point::new(t1.0, t1.1),
-                Point::new(t2.0, t2.1),
+                Point::new(t1.0 as isize, t1.1 as isize),
+                Point::new(t2.0 as isize, t2.1 as isize),
                 IndexedColor(16 - dis as u8),
             );
         }
@@ -23,9 +23,9 @@ fn _start() {
     window.wait_char();
 }
 
-const ADJUST_X: isize = 8;
-const ADJUST_Y: isize = 29;
-const TABLE: [(isize, isize); 16] = [
+const ADJUST_X: u8 = 8;
+const ADJUST_Y: u8 = 29;
+const TABLE: [(u8, u8); 16] = [
     (204 - ADJUST_X, 129 - ADJUST_Y),
     (195 - ADJUST_X, 90 - ADJUST_Y),
     (172 - ADJUST_X, 58 - ADJUST_Y),
